commit and push

+++ SchoolManagementSystem/src/main/java/com/school/repository/IStudentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.school.entity.Student;
 
-  public interface IStudentRepository extends JpaRepository<Student, Integer> {
+  public interface IStudentRepository extends JpaRepository<Student, Long> {
   }
